Move components into import-export folder
