+++ src/ggrc-client/js/components/import-export/templates/quick-tips.mustache
