+++ src/ggrc-client/js/entrypoints/import/index.js
@@ -7,7 +7,7 @@ import '../commons';
 
 // Components
 import '../../components/spinner/spinner';
-import '../../components/csv/import';
+import '../../components/import-export/import';
 import '../../components/object-list/object-list';
 import '../../components/tree_pagination/tree_pagination';
